Milestone 6/FinalCode-Team16/soft-engg-project-jan-2025-se-Jan-16-development/Backend/requirement.txt
aniso8601==10.0.0
blinker==1.9.0
click==8.1.8
Flask==3.1.0
Flask-Cors==5.0.0
Flask-JWT-Extended==4.7.1
Flask-RESTful==0.3.10
Flask-SQLAlchemy==3.1.1
greenlet==3.1.1
itsdangerous==2.2.0
Jinja2==3.1.5
MarkupSafe==3.0.2
PyJWT==2.10.1
pytz==2024.2
six==1.17.0
SQLAlchemy==2.0.37
typing_extensions==4.12.2
Werkzeug==3.1.3
streamlit
google-generativeai
python-dotenv
langchain
PyPDF2
chromadb
faiss-cpu
langchain_google_genai